[ISSUE-887][REFACTOR] Configuration type convert to Enum
