+++ client/src/test/java/org/apache/celeborn/client/ShuffleClientSuiteJ.java
@@ -39,11 +39,11 @@ import org.apache.commons.lang3.RandomStringUtils;
 import org.junit.Test;
 
 import org.apache.celeborn.client.compress.Compressor;
-import org.apache.celeborn.client.compress.Compressor.CompressionCodec;
 import org.apache.celeborn.common.CelebornConf;
 import org.apache.celeborn.common.identity.UserIdentifier;
 import org.apache.celeborn.common.network.client.TransportClient;
 import org.apache.celeborn.common.network.client.TransportClientFactory;
+import org.apache.celeborn.common.protocol.CompressionCodec;
 import org.apache.celeborn.common.protocol.PartitionLocation;
 import org.apache.celeborn.common.protocol.PbRegisterShuffleResponse;
 import org.apache.celeborn.common.protocol.message.ControlMessages.*;
