+++ worker/src/test/java/org/apache/celeborn/service/deploy/worker/storage/FileWriterSuiteJ.java
@@ -86,7 +86,7 @@ public class FileWriterSuiteJ {
   private static final CelebornConf CONF = new CelebornConf();
   public static final Long SPLIT_THRESHOLD = 256 * 1024 * 1024L;
   public static final PartitionSplitMode splitMode = PartitionSplitMode.HARD;
-  public static final PartitionType partitionType = PartitionType.REDUCE_PARTITION;
+  public static final PartitionType partitionType = PartitionType.REDUCE;
 
   private static File tempDir = null;
   private static LocalFlusher localFlusher = null;
